feat(briefing): extract news pipeline (commit 6/11)
Workstream 6 Step 1 — extracts the news pipeline from briefing-service.js
into a focused module + deletes TWO dead code paths along the way (Option A
precedent from commit 4 / claude_memory #294).

NEW: server/lib/briefing/pipelines/news.js
  - discoverNews({ snapshot, snapshotId }) — pipeline contract entry,
    returns { news: {items, reason}, reason }. Owns its writeSectionAndNotify
    and errorMarker .catch wrapper. Special case: news section content is
    itself { items, reason }, so the wider contract has nested reasons (inner
    section reason + outer orchestrator reason).
  - fetchRideshareNews({ snapshot }) — re-exported live path
    (Gemini 3 Pro + Google Search). Preserves the import surface for
    server/api/briefing/briefing.js (refresh endpoints) AND
    refreshNewsInBriefing in briefing-service.js (uses it via the same
    import binding).
  - filterRecentNews + buildNewsPrompt — module-private helpers used
    only by fetchRideshareNews.

DRIFT RESOLUTION: two dead news paths deleted (zero callers, same pattern
as fetchWeatherForecast in commit 4):
  - fetchNewsWithClaudeWebSearch (Claude WebSearch fallback, 86 lines):
    fetchRideshareNews does NOT fall back to it. Never wired up.
  - consolidateNewsItems (multi-provider deduplication helper, 60 lines):
    fetchRideshareNews does its own dedup via filterRecentNews + .sort()
    chain. Never called.
  - Resolution logged to claude_memory row #295 (parent_id=#294 — Commit 4's
    precedent row).

briefing-service.js delta (2,750 → 2,384, net -366 lines):
  - import { discoverNews, fetchRideshareNews } from pipelines/news.js
  - export { fetchRideshareNews } (re-export for external + internal callers)
  - orchestrator newsPromise: replaced .then/.catch chain with direct
    discoverNews() call (~10-line shrink, pipeline owns its own SSE write
    and errorMarker)
  - newsItems extraction: newsResult?.items → newsResult?.news?.items
    (line 2024); warning message updated to match
  - final-assembly: newsResult?.reason → newsResult?.news?.reason
    (line 2066, reads inner section reason)
  - refreshNewsInBriefing untouched: uses fetchRideshareNews directly via
    the re-export binding; flat {items, reason} shape unchanged
  - deleted 2 separate regions: lines 622-708 (Claude fallback, +JSDoc) and
    lines 1784-2064 (filterRecentNews JSDoc + 4 functions including
    consolidateNewsItems). 367 lines removed, 1 dead JSDoc collapsed.

Verification:
  - npm run typecheck (tsc --noEmit): clean
  - npm run lint (eslint client/src): clean
  - node --check on briefing-service.js, pipelines/news.js: pass
  - Dynamic import smoke: news.js exports { discoverNews, fetchRideshareNews }
  - Grep audits: zero orphan local definitions, zero literal channel strings,
    zero references to fetchNewsWithClaudeWebSearch or consolidateNewsItems
    outside intentional documentation comments
  - Re-export chain resolves: server/api/briefing/briefing.js still imports
    fetchRideshareNews from briefing-service.js (works via re-export)
  - Both error pathways preserved:
    Pathway A (thrown) → Promise.allSettled captures → newsResult undefined →
    failedReasons.news populated → 'news-fetch-failed' sentinel fires
    Pathway B (graceful) → discoverNews returns { news: {items, reason}, reason }
    → newsResult.news.items + newsResult.news.reason resolved at lines 2024/2066

Per PLAN_workstream6_remaining-2026-05-02.md §2.1 (wider contract),
§2.5 (re-export pattern), §2.3 (sed for line-range deletes).
Same Option A precedent as commit 4 (claude_memory #294).
